+++ src/Mzinga.Viewer/MessageHandlers.cs
@@ -13,9 +13,8 @@ using Avalonia.Controls.ApplicationLifetimes;
 
 using CommunityToolkit.Mvvm.Messaging;
 
-using Mzinga.Viewer;
+using Mzinga.Core;
 using Mzinga.Viewer.ViewModels;
-
 using Mzinga.Viewer.Views;
 
 namespace Mzinga.Viewer